Move S3 read_text functions to Ray data source
