+++ awswrangler/distributed/datasources/__init__.py
@@ -1,7 +1,9 @@
 """Distributed Datasources Module."""
 
+from awswrangler.distributed.datasources.pandas_datasource import PandasDatasource
 from awswrangler.distributed.datasources.parquet_datasource import ParquetDatasource
 
 __all__ = [
+    "PandasDatasource",
     "ParquetDatasource",
 ]
